requirements.txt
appdaemon>=4.4.0
aiohttp>=3.9
pyyaml>=6.0
voluptuous>=0.14.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
